fix(settings): Debounce filepicker calls
The filepicker falls apart when you open it twice in less than 200ms.
That is because the filepicker closes any previous instance when it
opens, and closing logic delayed by 200ms.

It happens like
1) Open file picker 1
2) Open file picker 2 -> trigger destruction of fp1, continue fp2
   initialization
3) fp2 destruction is finished, but the reference now points to fp1 ->
   fp1 gets destroyed

By adding debounce we can make sure focus+click do not cause a doubly
opened filepicker and the bug described above.

Signed-off-by: Christoph Wurst <christoph@winzerhof-wurst.at>
Signed-off-by: nextcloud-command <nextcloud-command@users.noreply.github.com>